Crate servo_base

A crate to hold very common types in Servo.

You should almost never need to add a data type to this crate. Instead look for a more shared crate that has fewer dependents.

Modules

cross_process_instant

An implementation of a monotonic, nanosecond precision timer, like std::time::Instant that can be serialized and compared across processes.

generic_channel

Enum wrappers to be able to select different channel implementations at runtime.

id

Namespaces and ids shared by many crates in Servo.

print_tree

rope 🔒

text

threadboost

Helper to boost critical threads

threadpool

unicode_block 🔒

Structs

Epoch

A struct for denoting the age of messages; prevents race conditions.

Rope

An implementation of a rope data structure, composed of lines of owned strings. This is used to implement text controls in Servo.

RopeChars

A Chars-like iterator for Rope.

RopeIndex

An index into a Rope data structure. Used to efficiently identify a particular position in a Rope. As Rope always uses Rust strings interally, code point indices represented in a RopeIndex are assumed to be UTF-8 code points (one byte each).

RopeSlice

A slice of a Rope. This can be used to to iterate over a subset of characters of a Rope or to return the content of the RopeSlice as a String.

Enums

RopeMovement

Describes a unit of movement for Rope::move_by.

Traits

IpcSend

Abstraction of the ability to send a particular type of message, used by net_traits::ResourceThreads to ease the use its IpcSender sub-fields XXX: If this trait will be used more in future, some auto derive might be appealing

WebRenderEpochToU16

Functions

read_json_from_file

write_json_to_file

Type Aliases

IpcSendResult