servo_base

Module generic_channel

Enum wrappers to be able to select different channel implementations at runtime.

Modules

buffered 🔒

callback 🔒

Generic Callbacks

generic_channelset 🔒

lazy_callback 🔒

Lazy Callbacks

oneshot 🔒

shared_memory 🔒

Structs

CallbackSetter

GenericBufferedSender

A buffered sender that collects individual messages (U) and sends them as a single batched message (T) via a user-provided packing closure.

GenericCallback

A mechanism to run a callback in the process this callback was constructed in.

GenericOneshotReceiver

The oneshot receiver struct

GenericOneshotSender

The oneshot sender struct

GenericReceiver

GenericReceiverSet

A GenericReceiverSet. Allows you to wait on multiple GenericReceivers. Automatically selects either Ipc or crossbeam depending on multiprocess mode.

GenericReceiverVisitor 🔒

GenericSender

A GenericSender that sends messages to a GenericReceiver.

GenericSenderVisitor 🔒

GenericSharedMemory

LazyCallback

Basic struct for LazyCallback

Enums

GenericReceiverVariants 🔒

GenericSelectionResult

Result for readable events returned from GenericReceiverSet::select.

GenericSenderVariants 🔒

The actual GenericSender variant.

ReceiveError

SendError

TryReceiveError

Statics

USE_IPC 🔒

Cache for being in Ipc Mode

Traits

GenericSend

Abstraction of the ability to send a particular type of message cross-process. This can be used to ease the use of GenericSender sub-fields.

Functions

channel

Creates a Servo channel that can select different channel implementations based on multiprocess mode or not. If the scenario doesn’t require message to pass process boundary, a simple crossbeam channel is preferred.

lazy_callback

A LazyCallback is a Callback that will be initialized at a later date. We return the ‘LazyCallback’ which is a GenericCallback. We also return a ‘CallbackSetter’ where the callback can be set at a later date.

new_generic_channel_crossbeam 🔒

Private helper function to create a crossbeam based channel.

new_generic_channel_ipc 🔒

oneshot

Creates a oneshot generic channel used to send only a single message, similar tokio::sync::oneshot. This is not the same as ipc_channel::oneshot. The send and receive methods will consume the Sender/Receiver. We will automatically select ipc or crossbeam channels.

serialize_generic_sender_variants 🔒

to_receive_result

use_ipc 🔒

Return if we should be in IPC Mode

Type Aliases

ReceiveResult

RoutedReceiver

RoutedReceiverReceiveResult

SendResult

TryReceiveResult