servo_base::generic_channel::callback

Type Alias MsgCallback

pub type MsgCallback<T> = dyn FnMut(Result<T, IpcError>) + Send;

The callback type of our messages.

This is equivalent to [TypedRouterHandler][ipc_channel::router::TypedRouterHandler], except that this type is not wrapped in a Box. The callback will be wrapped in either a Box or an Arc, depending on if it is run on the router, or passed to the recipient.