servo_base::generic_channel

Module generic_channelset

Structs

GenericReceiverSet

A GenericReceiverSet. Allows you to wait on multiple GenericReceivers. Automatically selects either Ipc or crossbeam depending on multiprocess mode.

Selector

Borrows of a GenericReceiverSet used to drive repeated select calls without rebuilding the underlying crossbeam_channel::Select each time. See GenericReceiverSet::selector.

Enums

GenericReceiverSetVariants 🔒

GenericSelectionResult

Result for readable events returned from GenericReceiverSet::select.

SelectorInner 🔒