servo_base::generic_channel

Module oneshot

Structs

GenericOneshotReceiver

The oneshot receiver struct

GenericOneshotSender

The oneshot sender struct

Functions

oneshot

Creates a oneshot generic channel used to send only a single message, similar tokio::sync::oneshot. This is not the same as ipc_channel::oneshot. The send and receive methods will consume the Sender/Receiver. We will automatically select ipc or crossbeam channels.