servo_base

Module id

Namespaces and ids shared by many crates in Servo.

Macros

namespace_id 🔒

size_of_test 🔒

Asserts the size of a type at compile time.

Structs

AtomicOptionScrollTreeNodeId

BlobIndex

BroadcastChannelRouterIndex

BrowsingContextGroupId

BrowsingContextIndex

CookieStoreIndex

CryptoKeyIndex

DomExceptionIndex

DomMatrixIndex

DomPointIndex

DomQuadIndex

DomRectIndex

FileIndex

FileListIndex

HistoryStateIndex

ImageBitmapIndex

ImageDataIndex

Index

A non-zero index, associated with a particular type.

MessagePortIndex

MessagePortRouterIndex

NamespaceIndex

A pipeline-namespaced index associated with a particular type.

OffscreenCanvasIndex

PainterId

PipelineIndex

PipelineNamespace

Each pipeline ID needs to be unique. However, it also needs to be possible to generate the pipeline ID from an iframe element (this simplifies a lot of other code that makes use of pipeline IDs).

PipelineNamespaceId

PipelineNamespaceInstaller

A per-process installer of pipeline-namespaces.

PipelineNamespaceRequest

Request a pipeline-namespace id from the constellation.

QuotaExceededErrorIndex

ScriptEventLoopId

ScrollTreeNodeId

An id for a ScrollTreeNode in the ScrollTree. This contains both the index to the node in the tree’s array of nodes as well as the corresponding SpatialId for the SpatialNode in the WebRender display list.

ServiceWorkerIndex

ServiceWorkerRegistrationIndex

WebViewId

ZeroIndex

An attempt to create a new Index value failed because the index value was zero.

Constants

CONSTELLATION_PIPELINE_NAMESPACE_ID

EMBEDDER_PIPELINE_NAMESPACE_ID

FIRST_CONTENT_PIPELINE_NAMESPACE_ID

The next available PipelineNamespaceId for the allocation in the constellation. Starting from 2, since we reserved namespace 0 for the embedder, and 1 for the constellation.

INSTALLED_SCRIPT_EVENT_LOOP_ID

PIPELINE_NAMESPACE

TEST_BROWSING_CONTEXT_ID

TEST_BROWSING_CONTEXT_INDEX

TEST_NAMESPACE

TEST_PAINTER_ID

TEST_PIPELINE_ID

TEST_PIPELINE_INDEX

TEST_SCRIPT_EVENT_LOOP_ID

TEST_WEBVIEW_ID

Statics

PAINTER_ID 🔒

PIPELINE_NAMESPACE_INSTALLER 🔒

A per-process unique pipeline-namespace-installer. Accessible via PipelineNamespace.

SCRIPT_EVENT_LOOP_ID 🔒

Traits

Indexable

A type that implements this trait is expected to be used as part of the NamespaceIndex type.

Type Aliases

BlobId

BroadcastChannelRouterId

BrowsingContextId

CookieStoreId

CryptoKeyId

DomExceptionId

DomMatrixId

DomPointId

DomQuadId

DomRectId

FileId

FileListId

HistoryStateId

ImageBitmapId

ImageDataId

MessagePortId

MessagePortRouterId

OffscreenCanvasId

PipelineId

QuotaExceededErrorId

ServiceWorkerId

ServiceWorkerRegistrationId