servo_base

Struct Rope

pub struct Rope {
    lines: Vec<String>,
}

An implementation of a rope data structure, composed of lines of owned strings. This is used to implement text controls in Servo.

Fields
lines: Vec<String>

The lines of the rope. Each line is an owned string that ends with a newline (\n), apart from the last line which has no trailing newline.

Implementations

impl Rope

pub fn new(contents: impl Into<String>) -> Self

pub fn contents(&self) -> String

pub fn first_index(&self) -> RopeIndex

pub fn last_index(&self) -> RopeIndex

pub fn replace_range( &mut self, range: Range<RopeIndex>, string: impl Into<String>, ) -> RopeIndex

Replace the given range of RopeIndexs with the given string. Returns the RopeIndex of the end of the insertion.

fn delete_range(&mut self, range: Range<RopeIndex>)

pub fn slice<'a>( &'a self, start: Option<RopeIndex>, end: Option<RopeIndex>, ) -> RopeSlice<'a>

Create a RopeSlice for this Rope from start to end. If either of these is None, then the slice will extend to the extent of the rope.

pub fn chars<'a>(&'a self) -> RopeChars<'a>

pub fn is_empty(&self) -> bool

Return true if the Rope is empty or false otherwise. This will also return true if the contents of the Rope are a single empty line.

pub fn len_utf16(&self) -> Utf16CodeUnits

The total number of code units required to encode the content in utf16.

fn line(&self, index: usize) -> &str

fn line_for_index(&self, index: RopeIndex) -> &String

fn line_for_index_mut(&mut self, index: RopeIndex) -> &mut String

fn last_index_in_line(&self, line: usize) -> RopeIndex

fn start_of_following_line(&self, index: RopeIndex) -> RopeIndex

Return a RopeIndex which points to the start of the subsequent line. If the given RopeIndex is already on the final line, this will return the final index of the entire Rope.

fn end_of_preceding_line(&self, index: RopeIndex) -> RopeIndex

Return a RopeIndex which points to the end of preceding line. If already at the end of the first line, this will return the start index of the entire Rope.

pub fn move_by( &self, origin: RopeIndex, unit: RopeMovement, amount: isize, ) -> RopeIndex

fn move_by_lines(&self, origin: RopeIndex, lines_to_move: isize) -> RopeIndex

fn move_by_iterator( &self, origin: RopeIndex, unit: RopeMovement, amount: isize, ) -> RopeIndex

pub fn normalize_index(&self, rope_index: RopeIndex) -> RopeIndex

Given a RopeIndex, clamp it and ensure that it is on a character boundary, meaning that its indices are all bound by the actual size of the line and the number of lines in this Rope.

pub fn index_to_utf8_offset(&self, rope_index: RopeIndex) -> Utf8CodeUnits

Convert a RopeIndex into a byte offset from the start of the content.

pub fn index_to_utf16_offset(&self, rope_index: RopeIndex) -> Utf16CodeUnits

pub fn index_to_character_offset(&self, rope_index: RopeIndex) -> Utf32CodeUnits

Convert a RopeIndex into a character offset from the start of the content.

pub fn utf8_offset_to_rope_index(&self, utf8_offset: Utf8CodeUnits) -> RopeIndex

Convert a byte offset from the start of the content into a RopeIndex.

pub fn utf16_offset_to_utf8_offset( &self, utf16_offset: Utf16CodeUnits, ) -> Utf8CodeUnits

pub fn relevant_word_boundaries<'a>(&'a self, index: RopeIndex) -> RopeSlice<'a>

Find the boundaries of the word most relevant to the given RopeIndex. Word returned in order or precedence:

If the index intersects the word or is the index directly preceding a word, the boundaries of that word are returned.

The word preceding the cursor.

If there is no word preceding the cursor, the start of the line to the end of the next word.

pub fn line_boundaries<'a>(&'a self, index: RopeIndex) -> RopeSlice<'a>

Return the boundaries of the line that contains the given RopeIndex.

fn character_at(&self, index: RopeIndex) -> Option<char>

fn character_before(&self, index: RopeIndex) -> Option<char>