servo_base::threadboost

Function boost_thread

pub fn boost_thread(priority: ThreadPriority, boost_affinity: BoostAffinity)

Hint to the scheduler that this thread should be prioritised.

No effect if pref!(perf_thread_boost_enabled) is false.

TODO: The exact API and inner-workings are subject to change:

This is a hint to servo / the embedder and can be a no-op.

We might want to pass a thread identifier (enum variant?) so that we (or the embedder) can customize the optimization based on the thread without relying on parsing the thread-name.

Some optimizations like thread affinity selection also affect children threads, if spawned after this call, so placement can be important.