servo_base::threadboost

Module linux_sysfs

Statics

NON_LITTLE_CPU_CORES 🔒

Functions

capacity_of 🔒

Per-cpu relative capacity (normalized to 1024 for the strongest core in a system)

non_little_cpus 🔒

Determine the CPU ids of cores not in the little class.

parse_cpu_list 🔒

pin_thread_to_medium_or_large_cpus 🔒

Try and pin this thread to cpu cores above the smallest capacity class.

Type Aliases

CoreId 🔒