servo_base

Trait IpcSend

pub trait IpcSend<T>
where
    T: Serialize + for<'de> Deserialize<'de>,
{
    // Required methods
    fn send(&self, _: T) -> IpcSendResult;
    fn sender(&self) -> IpcSender<T>;
}

Abstraction of the ability to send a particular type of message, used by net_traits::ResourceThreads to ease the use its IpcSender sub-fields XXX: If this trait will be used more in future, some auto derive might be appealing

Required Methods

fn send(&self, _: T) -> IpcSendResult

send message T

fn sender(&self) -> IpcSender<T>

get underlying sender